// amplify/graphql/API.ts
export type CreateConnectionInput = {
  connectionId: string,
  userId: string,
};

export type ModelConnectionReceivedConditionInput = {
  and?: Array< ModelConnectionReceivedConditionInput | null > | null,
  createdAt?: ModelStringInput | null,
  not?: ModelConnectionReceivedConditionInput | null,
  or?: Array< ModelConnectionReceivedConditionInput | null > | null,
  status?: ModelStringInput | null,
  updatedAt?: ModelStringInput | null,
};

export type CreateConnectionReceivedInput = {
  receiverId: string,
  senderId: string,
  status?: string | null,
};
// ...
export type CreateConnectionMutationVariables = {
  condition?: ModelConnectionConditionInput | null,
  input: CreateConnectionInput,
};

export type CreateConnectionMutation = {
  createConnection?:  {
    __typename: "Connection",
    connectionId: string,
    createdAt: string,
    updatedAt: string,
    user?:  {
      __typename: "User",
      aboutMe?: string | null,
      age?: number | null,
      createdAt: string,
      email: string,
      fullName?: string | null,
      gender?: string | null,
      id: string,
      interests?: Array< string | null > | null,
      profilePictureUrl?: string | null,
      race?: string | null,
      spokenLanguage?: string | null,
      status?: string | null,
      updatedAt: string,
    } | null,
    userId: string,
  } | null,
};

export type CreateConnectionReceivedMutationVariables = {
  condition?: ModelConnectionReceivedConditionInput | null,
  input: CreateConnectionReceivedInput,
};

export type CreateConnectionReceivedMutation = {
  createConnectionReceived?:  {
    __typename: "ConnectionReceived",
    createdAt: string,
    receiver?:  {
      __typename: "User",
      aboutMe?: string | null,
      age?: number | null,
      createdAt: string,
      email: string,
      fullName?: string | null,
      gender?: string | null,
      id: string,
      interests?: Array< string | null > | null,
      profilePictureUrl?: string | null,
      race?: string | null,
      spokenLanguage?: string | null,
      status?: string | null,
      updatedAt: string,
    } | null,
    receiverId: string,
    senderId: string,
    status?: string | null,
    updatedAt: string,
  } | null,
};

export type CreateConnectionRequestMutationVariables = {
  condition?: ModelConnectionRequestConditionInput | null,
  input: CreateConnectionRequestInput,
};

export type CreateConnectionRequestMutation = {
  createConnectionRequest?:  {
    __typename: "ConnectionRequest",
    createdAt: string,
    receiverId: string,
    sender?:  {
      __typename: "User",
      aboutMe?: string | null,
      age?: number | null,
      createdAt: string,
      email: string,
      fullName?: string | null,
      gender?: string | null,
      id: string,
      interests?: Array< string | null > | null,
      profilePictureUrl?: string | null,
      race?: string | null,
      spokenLanguage?: string | null,
      status?: string | null,
      updatedAt: string,
    } | null,
    senderId: string,
    status?: string | null,
    updatedAt: string,
  } | null,
};

export type CreateDropdownListMutationVariables = {
  condition?: ModelDropdownListConditionInput | null,
  input: CreateDropdownListInput,
};

export type CreateDropdownListMutation = {
  createDropdownList?:  {
    __typename: "DropdownList",
    createdAt: string,
    id: string,
    name: string,
    options?: Array< string | null > | null,
    updatedAt: string,
  } | null,
};

export type CreateMessageMutationVariables = {
  condition?: ModelMessageConditionInput | null,
  input: CreateMessageInput,
};

export type CreateMessageMutation = {
  createMessage?:  {
    __typename: "Message",
    content?: string | null,
    createdAt: string,
    id: string,
    room?:  {
      __typename: "Room",
      createdAt: string,
      id: string,
      updatedAt: string,
    } | null,
    roomId: string,
    senderId: string,
    timestamp: string,
    translatedContent?: string | null,
    updatedAt: string,
  } | null,
};

export type CreateRaceListMutationVariables = {
  condition?: ModelRaceListConditionInput | null,
  input: CreateRaceListInput,
};

export type CreateRaceListMutation = {
  createRaceList?:  {
    __typename: "RaceList",
    createdAt: string,
    id: string,
    name: string,
    updatedAt: string,
  } | null,
};

export type CreateRoomMutationVariables = {
  condition?: ModelRoomConditionInput | null,
  input: CreateRoomInput,
};

export type CreateRoomMutation = {
  createRoom?:  {
    __typename: "Room",
    createdAt: string,
    id: string,
    messages?:  {
      __typename: "ModelMessageConnection",
      nextToken?: string | null,
    } | null,
    roomUsers?:  {
      __typename: "ModelRoomUserConnection",
      nextToken?: string | null,
    } | null,
    updatedAt: string,
  } | null,
};

export type CreateRoomUserMutationVariables = {
  condition?: ModelRoomUserConditionInput | null,
  input: CreateRoomUserInput,
};

export type CreateRoomUserMutation = {
  createRoomUser?:  {
    __typename: "RoomUser",
    createdAt: string,
    id: string,
    room?:  {
      __typename: "Room",
      createdAt: string,
      id: string,
      updatedAt: string,
    } | null,
    roomId: string,
    updatedAt: string,
    user?:  {
      __typename: "User",
      aboutMe?: string | null,
      age?: number | null,
      createdAt: string,
      email: string,
      fullName?: string | null,
      gender?: string | null,
      id: string,
      interests?: Array< string | null > | null,
      profilePictureUrl?: string | null,
      race?: string | null,
      spokenLanguage?: string | null,
      status?: string | null,
      updatedAt: string,
    } | null,
    userId: string,
  } | null,
};

export type CreateUserMutationVariables = {
  condition?: ModelUserConditionInput | null,
  input: CreateUserInput,
};

export type CreateUserMutation = {
  createUser?:  {
    __typename: "User",
    aboutMe?: string | null,
    age?: number | null,
    connectionReceived?:  {
      __typename: "ModelConnectionReceivedConnection",
      nextToken?: string | null,
    } | null,
    connectionRequests?:  {
      __typename: "ModelConnectionRequestConnection",
      nextToken?: string | null,
    } | null,
    connections?:  {
      __typename: "ModelConnectionConnection",
      nextToken?: string | null,
    } | null,
    createdAt: string,
    email: string,
    fullName?: string | null,
    gender?: string | null,
    id: string,
    interests?: Array< string | null > | null,
    profilePictureUrl?: string | null,
    race?: string | null,
    rooms?:  {
      __typename: "ModelRoomUserConnection",
      nextToken?: string | null,
    } | null,
    spokenLanguage?: string | null,
    status?: string | null,
    updatedAt: string,
  } | null,
};

export type DeleteConnectionMutationVariables = {
  condition?: ModelConnectionConditionInput | null,
  input: DeleteConnectionInput,
};

export type DeleteConnectionMutation = {
  deleteConnection?:  {
    __typename: "Connection",
    connectionId: string,
    createdAt: string,
    updatedAt: string,
    user?:  {
      __typename: "User",
      aboutMe?: string | null,
      age?: number | null,
      createdAt: string,
      email: string,
      fullName?: string | null,
      gender?: string | null,
      id: string,
      interests?: Array< string | null > | null,
      profilePictureUrl?: string | null,
      race?: string | null,
      spokenLanguage?: string | null,
      status?: string | null,
      updatedAt: string,
    } | null,
    userId: string,
  } | null,
};

export type DeleteConnectionReceivedMutationVariables = {
  condition?: ModelConnectionReceivedConditionInput | null,
  input: DeleteConnectionReceivedInput,
};

export type DeleteConnectionReceivedMutation = {
  deleteConnectionReceived?:  {
    __typename: "ConnectionReceived",
    createdAt: string,
    receiver?:  {
      __typename: "User",
      aboutMe?: string | null,
      age?: number | null,
      createdAt: string,
      email: string,
      fullName?: string | null,
      gender?: string | null,
      id: string,
      interests?: Array< string | null > | null,
      profilePictureUrl?: string | null,
      race?: string | null,
      spokenLanguage?: string | null,
      status?: string | null,
      updatedAt: string,
    } | null,
    receiverId: string,
    senderId: string,
    status?: string | null,
    updatedAt: string,
  } | null,
};

export type DeleteConnectionRequestMutationVariables = {
  condition?: ModelConnectionRequestConditionInput | null,
  input: DeleteConnectionRequestInput,
};

export type DeleteConnectionRequestMutation = {
  deleteConnectionRequest?:  {
    __typename: "ConnectionRequest",
    createdAt: string,
    receiverId: string,
    sender?:  {
      __typename: "User",
      aboutMe?: string | null,
      age?: number | null,
      createdAt: string,
      email: string,
      fullName?: string | null,
      gender?: string | null,
      id: string,
      interests?: Array< string | null > | null,
      profilePictureUrl?: string | null,
      race?: string | null,
      spokenLanguage?: string | null,
      status?: string | null,
      updatedAt: string,
    } | null,
    senderId: string,
    status?: string | null,
    updatedAt: string,
  } | null,
};

export type DeleteDropdownListMutationVariables = {
  condition?: ModelDropdownListConditionInput | null,
  input: DeleteDropdownListInput,
};

export type DeleteDropdownListMutation = {
  deleteDropdownList?:  {
    __typename: "DropdownList",
    createdAt: string,
    id: string,
    name: string,
    options?: Array< string | null > | null,
    updatedAt: string,
  } | null,
};

export type DeleteMessageMutationVariables = {
  condition?: ModelMessageConditionInput | null,
  input: DeleteMessageInput,
};

export type DeleteMessageMutation = {
  deleteMessage?:  {
    __typename: "Message",
    content?: string | null,
    createdAt: string,
    id: string,
    room?:  {
      __typename: "Room",
      createdAt: string,
      id: string,
      updatedAt: string,
    } | null,
    roomId: string,
    senderId: string,
    timestamp: string,
    translatedContent?: string | null,
    updatedAt: string,
  } | null,
};

export type DeleteRaceListMutationVariables = {
  condition?: ModelRaceListConditionInput | null,
  input: DeleteRaceListInput,
};

export type DeleteRaceListMutation = {
  deleteRaceList?:  {
    __typename: "RaceList",
    createdAt: string,
    id: string,
    name: string,
    updatedAt: string,
  } | null,
};

export type DeleteRoomMutationVariables = {
  condition?: ModelRoomConditionInput | null,
  input: DeleteRoomInput,
};

export type DeleteRoomMutation = {
  deleteRoom?:  {
    __typename: "Room",
    createdAt: string,
    id: string,
    messages?:  {
      __typename: "ModelMessageConnection",
      nextToken?: string | null,
    } | null,
    roomUsers?:  {
      __typename: "ModelRoomUserConnection",
      nextToken?: string | null,
    } | null,
    updatedAt: string,
  } | null,
};

export type DeleteRoomUserMutationVariables = {
  condition?: ModelRoomUserConditionInput | null,
  input: DeleteRoomUserInput,
};

export type DeleteRoomUserMutation = {
  deleteRoomUser?:  {
    __typename: "RoomUser",
    createdAt: string,
    id: string,
    room?:  {
      __typename: "Room",
      createdAt: string,
      id: string,
      updatedAt: string,
    } | null,
    roomId: string,
    updatedAt: string,
    user?:  {
      __typename: "User",
      aboutMe?: string | null,
      age?: number | null,
      createdAt: string,
      email: string,
      fullName?: string | null,
      gender?: string | null,
      id: string,
      interests?: Array< string | null > | null,
      profilePictureUrl?: string | null,
      race?: string | null,
      spokenLanguage?: string | null,
      status?: string | null,
      updatedAt: string,
    } | null,
    userId: string,
  } | null,
};

export type DeleteUserMutationVariables = {
  condition?: ModelUserConditionInput | null,
  input: DeleteUserInput,
};
// ...
export type UpdateConnectionMutationVariables = {
  condition?: ModelConnectionConditionInput | null,
  input: UpdateConnectionInput,
};
// ...
export type UpdateConnectionReceivedMutationVariables = {
  condition?: ModelConnectionReceivedConditionInput | null,
  input: UpdateConnectionReceivedInput,
};